Updated position for mug generation (center square). MockupPositionItem do not round positions if they're given as float values.

+++ src/Structures/Generator/MockupPositionItem.php
@@ -54,12 +54,12 @@ class MockupPositionItem
         }
 
         return [
-            'area_width' => (int)$this->areaWidth,
-            'area_height' => (int)$this->areaHeight,
-            'width' => (int)$this->width,
-            'height' => (int)$this->height,
-            'top' => (int)$this->top,
-            'left' => (int)$this->left,
+            'area_width' => (float)$this->areaWidth,
+            'area_height' => (float)$this->areaHeight,
+            'width' => (float)$this->width,
+            'height' => (float)$this->height,
+            'top' => (float)$this->top,
+            'left' => (float)$this->left,
         ];
     }
 }
